store and pass nowPlaying times and gif creation times as floats to allow ms persicion;

+++ src/avconvutil.h
@@ -5,7 +5,7 @@
 
 namespace avconfutil {
 
-QString time(int second);
+QString time(float second);
 QString resolution(int width, int height);
 
 } // namespace
